The display(legend name) now can be changed
in former version, we cannnot change the filename in list view,
because of Model.logfiles key is not modified.
in this version, this bug has been fixed.
the move-up/down buttons are hown, but this does not work.
the check box in the list-view also does not work.
TODO: implement these buttons and check-box
check-box will work as 'plot-enabled' button